requirements.txt
python-dotenv==1.0.1
motor==3.5.1
python-multipart==0.0.9
pymongo==4.8.0
boto3==1.35.14
langgraph==0.2.22
langchain-openai==0.1.23
streamlit==1.38.0
langchain-core==0.2.43